[TaskLocal] Remove comment which is true, but is confusing to end users
because it discusses an implementation detail of the type the comment is
on. Instead, when developers read comments they think about what they
can store inside a task local -- so the comment about "task local must
be a class" led them to believe THEIR types must be classes, while the
comment was explaining the implementation detail of `class
TaskLocal<T>`.

Resolves rdar://117982608