[Playgrounds] Restore the `performPlaygroundTransform()` variant that takes a high-perf flag
Temporarily restore the old variant of the `performPlaygroundTransform()` that takes a boolean to determine whether to only apply high-performance playground transforms. This function is still used by LLDB and possibly other clients, and should continue to work. It can be removed once all clients have switched over to the more generic variant that takes a `PlaygroundOptionSet`.

Also added a comment to the old function indicating that new code should use the more general form.