[ClangImporter] Always rebuild any bridging PCH that has errors
If a PCH was output with errors, it may not have serialized all its
inputs. If there was a change to the search path or a headermap now
exists where it didn't previously, it's possible those inputs will now
be found. Ideally we would only rebuild in this particular case rather
than any error in general, but explicit module builds are the real
solution there. For now, just treat PCH with errors as out of date.

Resolves rdar://117037471.

(cherry picked from commit 322b03d0f93899bae3d27c8e241cd5737d46e749)