[Swift-VFE] Prevent VFE on type hierarchies split across modules
The included test demonstrates the issues most clearly, but in short
when a type hierarchy is split across two modules, you cannot guarantee
that the `llvm.type.check.load` instructions will be visible to
GlobalDCE as virtual calls can be made using a dispatch thunk
located in a different link unit.

This change detects when the type hierarchies is split between modules
and opts out of VFE. Fortunately this is disallowed by default,
requiring the open access modifier. I hope this case is relatively
uncommon as a result.