[swift-wme] Prevent WME on cross module conformances
If a type conforms to a protocol defined in a different module
(and therefore potentially a different LTO unit), the protocol dispatch
thunk and protocol witness table can end up in different LTO units,
making WME unsafe.

Avoid this by only applying vcall visibility to the witness table
if all protocols conformed to are in the same module as the conforming
type.