[swift-vfe] Eliminate unused base method implementations
The LLVM VFE infrastructure was built to enable eliminating
virtual functions that cannot possibly be called. Previously,
given methods Base.foo and Derived.foo, we were only able to
eliminate either if neither Base.foo nor Derived.foo are called.

This patch improves this in that if all calls to foo are made
through Derived, we can eliminate Base.foo. It only does this
if all calls to foo are made within the defining module for now.

It is possible to expand this to inter-module vcalls, but more
complicated. (We need our own thunks instead of re-using the reslient
dispatch thunks)

This brings us closer to the C++ implementation of VFE.