AST/Sema: Refactor lazy TypeRefinementContext expansion.
When building out the TypeRefinementContext tree, if we encounter an AST node
that we cannot fully build out the children for, instead add a leaf node to the
tree and mark it as needing expansion. Then skip walking the children of the
AST node. When querying the tree, if when a node needing expansion is
encountered, expand the children of the node on-demand. The previous
implementation of lazy tree expansion did not actually skip the children of AST
nodes requiring expansion, so the resulting TypeRefinementContext tree could
contain malformed nodes with duplicate descendants.