[build] Add ASTGen libraries via a CMake function
We currently add all modules in `lib/ASTGen` via the top-level `CMakeLists.txt` and then modify them. in a loop at the end of the file. This doesn’t scale as we add more modules.

Create a CMake function that adds a Swift compiler library and move the definition of these libraries to separate `CMakeLists.txt` files in `Sources` subdirectories.