api-digester: Don’t compare invalid attributes
The API digester represents unknown attributes as DAK_Count, but passing this value to the `is{Adding,Removing}Breaking{API,ABI}()` functions is not permitted. On some platforms this would cause a crash; on others it hit unreachable code which happened to end up working.

Skip these unknown attributes during comparisons to avoid this problem.