api-digester: Support old name for @backDeployed
Swift 5.7’s API digester supported the unofficial @_backDeploy attribute, writing it to digests under the name “BackDeploy”, but by the time it the proposal was accepted, the spelling in source code was @backDeployed and the spelling in digests became “BackDeployed”. Add a special case to the new digester to accept the old spelling. This should fix an issue seen in Swift CI where the WholeSDK API digester test couldn’t process the StoreKit framework.