[5.10] [Profiler] Remove an assertion
Due to the recent change to how we simplify
counters (rdar://118185163), we can now end up
hitting this assertion for certain do-catch
statements. The counter has always been wrong
here, and previously would have underflowed at
runtime (rdar://100470244). This issue has been
properly fixed on main, but we haven't
cherry-picked that fix to 5.10 since it requires a
bunch of other changes. For now, just remove the
assertion to maintain the previous behavior.

rdar://118896974