[Runtime] Assert the correct name when retrieving standard Concurrency type descriptors.
If the runtime and Concurrency runtime are mismatched and have different, we can end up grabbing the wrong descriptor when looking up a standard Concurrency type. If they have different declarations of ConcurrencyStandardTypeDescriptors then the fields can be misaligned, causing us to retrieve the wrong one. This can cause extremely confusing type lookup failures or other crashes.

To better detect this problem in asserts builds, add an assert that the descriptor we got from concurrencyDescriptors actually has the right name.